Remove username and channel
We cannot use trassir/stable for packages
from conan-center-index, as username and
channel has become optional in conan, and
conan-center-index maintainers has decided
to remove them from dependencies' description
in recipes.

Recipes without username/channel are converted
into '_/_' (username/channel) and are not
found.

E.g. we have zlib/1.2.11@trassir/stable in our bintray
but openssl recipe requires zlib/1.2.11, so it searches
for zlib/1.2.11@_/_.

It's not an option to fix every recipe we need
in this monorepository, so let's adhere to their
conversion.

We still can use username/channel for our recipes
if there is such a need.